% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_classify.R
\name{as_tibble.inla.posterior.sample}
\alias{as_tibble.inla.posterior.sample}
\title{Convert posterior sample to tibble long format}
\usage{
\method{as_tibble}{inla.posterior.sample}(obj, var = c("hyperpar", "latent", "logdens"))
}
\arguments{
\item{obj}{return obj from \code{\link[inlar]{inla.posterior.sample}}.}
}
\description{
Convert posterior sample to tibble long format
}
