% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlar.R
\name{as_tibble.inla}
\alias{as_tibble.inla}
\title{Convert posterior sample to tibble long format}
\usage{
\method{as_tibble}{inla}(
  obj,
  type = char(summary, marginals),
  term = char(random, linear.predictor, fitted.values)
)
}
\arguments{
\item{obj}{return obj from \code{\link[inlar]{inla.posterior.sample}}.}

\item{type}{summary or marginals>}

\item{term}{which of the random, linear.predictor, or fitted.values}
}
\description{
Convert posterior sample to tibble long format
}
