% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlagg.R
\name{autoplot.inla}
\alias{autoplot.inla}
\title{Autoplot INLA obj}
\usage{
\method{autoplot}{inla}(obj, type = char(summary), term = char(random, fitted.values))
}
\arguments{
\item{obj}{INLA model}

\item{type}{summary or marginal}

\item{term}{term to plot}

\item{plot}{default to FALSE retrun a tibble available on assign.}
}
\description{
With \code{\link[inlar]{as_tibble}}, plotting the result is straightforward
with ggplot, here is just some basic quick plots which might not be what you
want.
}
