% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlar.R
\name{inla.posterior.sample}
\alias{inla.posterior.sample}
\title{Wrapper to add class to INLA's output}
\usage{
inla.posterior.sample(...)
}
\arguments{
\item{...}{Additional arguments to \code{fun}}
}
\description{
See \code{\link[INLA]{inla.posterior.sample}}
}
